refactor(api): serve profile names from /api/profiles/names
The endpoint has no council coupling and splitter admin tables need the
same lookup. The old path still exports the handler.
