+++ src/app/flow-councils/membership/VoterTable.tsx
@@ -230,7 +230,7 @@ export default function VoterTable(props: VoterTableProps) {
 
       for (const batch of splitIntoChunks(addresses, PROFILE_BATCH)) {
         try {
-          const res = await fetch("/api/flow-council/voter-groups/profiles", {
+          const res = await fetch("/api/profiles/names", {
             method: "POST",
             headers: { "Content-Type": "application/json" },
             body: JSON.stringify({ addresses: batch }),
