+++ src/app/flow-councils/permissions/permissions.tsx
@@ -293,7 +293,7 @@ export default function Permissions(props: PermissionsProps) {
 
       for (const batch of splitIntoChunks(managerAddresses, PROFILE_BATCH)) {
         try {
-          const res = await fetch("/api/flow-council/voter-groups/profiles", {
+          const res = await fetch("/api/profiles/names", {
             method: "POST",
             headers: { "Content-Type": "application/json" },
             body: JSON.stringify({ addresses: batch }),
